% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{anvil_logo}
\alias{anvil_logo}
\title{Add AnVIL logo to plot}
\usage{
anvil_logo(
  x_unit = 25,
  y_unit = 10,
  height_unit = 12,
  width_unit = 36,
  full_logo = TRUE
)
}
\arguments{
\item{x_unit}{Numeric value indiciating the amount of horizontal shifting
Higher values shift towards the right, lower values shift towards the left}

\item{y_unit}{Numeric value indicating the amount of vertical shifting.
Greater values result in downward shifting, smaller values result in an upward shifting}

\item{height_unit}{Numeric value specifying height}

\item{width_unit}{Numeric value specifying width}

\item{full_logo}{Bool; default is TRUE. If TRUE, will use the full logo; if FALSE, will use the image of an anvil only}
}
\value{
a list with ggplot2 components: \code{annotation_custom()}, \code{coord_cartesian()}, \code{theme()}
}
\description{
Logo placed at the bottom right corner of a ggplot
}
\details{
Adapted from
https://www.markhw.com/blog/logos
https://stackoverflow.com/q/76354700/14804653
https://github.com/FredHutch/hutchplot
}
\examples{
\dontrun{
library(ggplot2)

ggplot(diamonds[sample(nrow(diamonds), 1000), ], aes(carat, price)) +
geom_point(aes(colour = color)) +
scale_colour_palette_d(anvil_palette_sequential$yellow) +
anvil_logo(x_unit = 25, y_unit = 10, width_unit = 18, full_logo = FALSE)
}

}
