% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anvil_palette.R
\docType{data}
\name{anvil_palettes}
\alias{anvil_palettes}
\alias{anvil_palette_discrete}
\alias{anvil_palette_sequential}
\alias{anvil_palette_sequential_blues}
\alias{anvil_palette_diverging}
\alias{anvil_palette_full}
\alias{anvil_palette_contrast2}
\alias{anvil_palette_complement2}
\alias{anvil_palette_single}
\title{The Analysis Visualization and Informatics Lab-space (AnVIL) Palettes}
\format{
\subsection{\code{anvil_palettes}}{

An object of class \code{palettes_palette} with 1
colour palettes. Use \code{names("anvil_palettes")} to return all palette names.
}

\subsection{\code{anvil_palette_discrete}}{

An object of class \code{palettes_palette} with
2 discrete colour palettes.
Use \code{names(anvil_palette_discrete)} to return all palette names.
}

\subsection{\code{anvil_palette_sequential}}{

An object of class \code{palettes_palette} with
3 sequential colour palettes.
Use \code{names(anvil_palette_sequential)} to return all palette names.
}

\subsection{\code{anvil_palette_sequential_blues}}{

An object of class \code{palettes_palette} with
1 sequential colour palettes.
Use \code{names(anvil_palette_sequential_blues)} to return all palette names.
}

\subsection{\code{anvil_palette_diverging}}{

An object of class \code{palettes_palette} with
1 diverging colour palettes.
Use \code{names(anvil_palette_diverging)}  to return all palette names.
}

\subsection{\code{anvil_palette_full}}{

An object of class \code{palettes_palette} with
1 colour palettes.
Use \code{names(anvil_palette_full)} to return all palette names.
}

\subsection{\code{anvil_palette_contrast2}}{

An object of class \code{palettes_palette} with
1 contrasting colour palettes.
Use \code{names(anvil_palette_contrast2)} to return all palette names.
}

\subsection{\code{anvil_palette_complement2}}{

An object of class \code{palettes_palette} with
1complementing colour palettes.
Use \code{names(anvil_palette_complement2)} to return all palette names.
}

\subsection{\code{anvil_palette_single}}{

An object of class \code{palettes_palette} with
1 colour palettes.
Use \code{names(anvil_palette_single)} to return all palette names.
}
}
\source{
\url{https://github.com/fhdsl/AnVILplot}
}
\usage{
anvil_palettes

anvil_palette_discrete

anvil_palette_sequential

anvil_palette_sequential_blues

anvil_palette_diverging

anvil_palette_full

anvil_palette_contrast2

anvil_palette_complement2

anvil_palette_single
}
\description{
Palettes inspired by the official colors of the AnVIL
}
\examples{
# Get all palettes by name.
names(anvil_palettes)

# Plot all palettes.
plot(anvil_palettes)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_colour]{pal_colour()}}
}
\author{
\href{https://github.com/kweav}{Kate Isaac}
}
\keyword{datasets}
