% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVILplot-package.R
\docType{package}
\name{AnVILplot-package}
\alias{AnVILplot}
\alias{AnVILplot-package}
\title{AnVILplot: AnVIL Colour Palettes}
\description{
Colour palettes inspired by the AnVIL logo. Built using the 'palettes' package, which provides methods for printing, formatting, casting and coercion, extraction and updating of components, plotting, colour mixing arithmetic, and colour interpolation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fhdsl/AnVILplot}
  \item Report bugs at \url{https://github.com/fhdsl/AnVILplot/issues}
}

}
\author{
\strong{Maintainer}: Kate Isaac \email{kisaac@fredhutch.org}

}
\keyword{internal}
